core/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-docker-helper==1.0.0.1
django-filter==25.1
djangorestframework==3.16.0
drf-generators==0.5.0
gunicorn==23.0.0
Markdown==3.8.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
asgiref==3.9.1
Django==5.2.4
django-docker-helper==1.0.0.1
django-filter==25.1
django-jazzmin==3.0.1
django-viewflow==2.2.12
djangorestframework==3.16.0
drf-generators==0.5.0
gunicorn==23.0.0
Markdown==3.8.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
asgiref==3.9.1
Django==5.2.4
django-docker-helper==1.0.0.1
django-filter==25.1
django-jazzmin==3.0.1
django-theme-soft-design==1.0.11
django-viewflow==2.2.12
djangorestframework==3.16.0
drf-generators==0.5.0
gunicorn==23.0.0
Markdown==3.8.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
